Add list file for refactor
