+++ book.rb
@@ -7,8 +7,7 @@ class Book
     @rentals = []
   end
 
-  def add_rental(rental)
-    @rentals.push(rental)
-    rental.book = self
+  def add_rental(date, person)
+    Rental.new(date, self, person)
   end
 end
